+++ app/Controllers/MailLogController.php
@@ -362,7 +362,7 @@ class MailLogController extends ViewController
 			'auth_provider' => $this->session->get('auth_provider'),
 			'current_route' => $this->request->getPathInfo(),
 			'rspamd_stats' => $this->getRspamdStat(),
-			'quarantine_days' => $_ENV['QUARANTINE_DAYS'],
+			'quarantine_days' => (int) $_ENV['QUARANTINE_DAYS'],
 		]));
 	}