+++ app/Utils/Helper.php
@@ -276,11 +276,11 @@ class Helper {
 	// Snippet from PHP Share: http://www.phpshare.org
 	public static function formatSizeUnits($bytes) {
 	if ($bytes >= 1073741824) {
-	    $bytes = number_format($bytes / 1073741824, 1) . 'G';
+	    $bytes = number_format($bytes / 1073741824, 1, ',', '.') . 'G';
 	} elseif ($bytes >= 1048576) {
-	    $bytes = number_format($bytes / 1048576, 1) . 'M';
+	    $bytes = number_format($bytes / 1048576, 1, ',', '.') . 'M';
 	} elseif ($bytes >= 1024) {
-	    $bytes = number_format($bytes / 1024, 1) . 'K';
+	    $bytes = number_format($bytes / 1024, 1, ',', '.') . 'K';
 	} elseif ($bytes > 1) {
 	    $bytes = $bytes . 'B';
 	} elseif ($bytes == 1) {