+++ app/Forms/MailAliasForm.php
@@ -82,7 +82,7 @@ class MailAliasForm extends AbstractType
 	public static function create(
 			FormFactoryInterface $formFactory,
 			Request $request,
-			$data = null): Form {
+			array $data = null): Form {
 
 		return FormHelper::formCreator($formFactory, $request, self::class, $data);
 	}