+++ app/Forms/MailReleaseForm.php
@@ -78,7 +78,7 @@ class MailReleaseForm extends AbstractType
 	public static function create(
 			FormFactoryInterface $formFactory,
 			Request $request,
-			$data = null,
+			array $data = null,
 			array $options = []): Form {
 
 		return FormHelper::formCreator($formFactory, $request, self::class, $data, $options);