+++ app/Forms/MapWithCustomFieldForm.php
@@ -92,7 +92,7 @@ class MapWithCustomFieldForm extends AbstractType
 	public static function create(
 			FormFactoryInterface $formFactory,
 			Request $request,
-			$data = null,
+			array $data = null,
 			array $options = []): Form {
 
 		// Merge class with any existing class