+++ app/Forms/MapWithOneFieldForm.php
@@ -60,7 +60,7 @@ class MapWithOneFieldForm extends AbstractType
 	public static function create(
 			FormFactoryInterface $formFactory,
 			Request $request,
-			$data = null,
+			array $data = null,
 			array $options = []): Form {
 
 		// Merge class with any existing class