+++ app/Forms/MapWithTwoFieldsForm.php
@@ -63,7 +63,7 @@ class MapWithTwoFieldsForm extends AbstractType
 	public static function create(
 			FormFactoryInterface $formFactory,
 			Request $request,
-			$data = null,
+			array $data = null,
 			array $options = []): Form {
 
 		// Merge class with any existing class