+++ app/Forms/UserForm.php
@@ -159,7 +159,7 @@ class UserForm extends AbstractType
 	public static function create(
 			FormFactoryInterface $formFactory,
 			Request $request,
-			$data = null,
+			array $data = null,
 			array $options = []): Form {
 
 		return FormHelper::formCreator($formFactory, $request, self::class, $data, $options);